requirements.txt
tensorflow>=2.0
keras>=2.3.0
Pillow>=8.0.0
numpy>=1.18.0
scikit-learn>=0.22